More neurons
Added 81 more neurons in the first hidden layer and 54 more in the
second hidden layer. Can't figure out why after the first time you input
in the GUI the results come out fine, but after the second input it out
put the same results as from the first input even if the pattern is
different, it will only change if you enter a blank input.