Added some stuff
Fixed MNIST to txt.py to output correctly, 30 hidden.pas and Готовая.pas
are pretty much the same thing just with a few tweaks and differences.
Готовая.pas is the final version(sort of). There is a bug that when you
input with the GUI for the first time it works, then when you try for
the second time it will output the same result unless you restart the
program. There is also another bug where the more times you train the
program the less accurate it is, for example: if you train it for around
1000 times it will get it right every 1 to 3 time, but if you train for
several thousands of times it doesn't gets it right maybe once every 10
to 15 times. Sort of counterintuitive if I say so myself.